Sove issue 10: New translation rule is added for event synchronisation independent of machine inclusion.
